src/requirements.txt
psycopg2-binary==2.9.10
typing_extensions==4.12.2
